Transfered client cabinet to client component
