--- src/app/modules/client/client-cabinet/client-cabinet.component.css
@@ -1,16 +1,0 @@
-app-animal-item {
-  display: inline-block;
-  margin: 6px;
-}
-app-create-animal-item {
-  margin: 6px;
-  position: relative;
-  display: inline-block;
-  flex-direction: column;
-}
-#appointments-div {
-  margin-top: 30px;
-}
-#appointments-div h2 {
-  text-align: center;
-}
