+++ src/app/modules/client/client-edit-form/client-edit-form.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { ApiService } from 'src/app/services/api.service';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { AbstractControl, FormControl, FormGroup, ValidatorFn, Validators } from '@angular/forms';
 import { AuthService } from 'src/app/services/auth.service';
 import { ClientService } from 'src/app/services/client.service' 
 import { EmailService} from 'src/app/services/email.service' 
@@ -31,10 +31,7 @@ export class ClientEditFormComponent implements OnInit {
     ]),
     phoneNumber: new FormControl('',[
       Validators.required, Validators.pattern("^[0-9]{12}$")
-    ]),
-    password: new FormControl('',[
-      Validators.required, Validators.pattern("^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{8,}$")
-    ]),
+    ])
     })
     constructor(public apiService: ApiService, public http: HttpClient, public authService: AuthService,
       public clientService: ClientService, private emailService: EmailService) {   
@@ -45,10 +42,8 @@ export class ClientEditFormComponent implements OnInit {
   }
   
     isControlInvalid(controlName: string): boolean {
-      const control = this.clientForm.controls[controlName];
-      
-      const result = control.invalid && control.touched;
-      
+      const control = this.clientForm.controls[controlName];      
+      const result = control.invalid && control.touched;      
       return result;
   }
   
@@ -61,7 +56,6 @@ export class ClientEditFormComponent implements OnInit {
             return;
       }   
       else {
-        alert("input successful")
         this.edit();
       }
      
@@ -75,7 +69,6 @@ export class ClientEditFormComponent implements OnInit {
         "lastName": this.clientForm.value.lastName,
         "email": this.clientForm.value.email,
         "phoneNumber": this.clientForm.value.phoneNumber,
-        "password": this.clientForm.value.password,
         "userId" : this.id
       }
       this.apiService.updateEntity('clients', this.client.id, editedClient).subscribe(
