+++ src/app/modules/client/client.component.css
@@ -1,0 +1,17 @@
+app-animal-item {
+    display: inline-block;
+    margin: 6px;
+  }
+  app-create-animal-item {
+    margin: 6px;
+    position: relative;
+    display: inline-block;
+    flex-direction: column;
+  }
+  #appointments-div {
+    margin-top: 30px;
+  }
+  #appointments-div h2 {
+    text-align: center;
+  }
+
